+++ prospector/tools/base.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 
 
 class ToolBase(object):
@@ -13,14 +14,19 @@ class ToolBase(object):
         its own configuration settings unless this method discovers some
         tool-specific configuration that should be used instead.
 
-        :return: A string indicating how or where this tool was configured from.
-                 For example, this can be a path to the .pylintrc file used, if
-                 used. Returning None means that prospector defaults were used.
+        :return: A tuple: the first element is a string indicating how or where
+                 this tool was configured from. For example, this can be a path
+                 to the .pylintrc file used, if used. None means that prospector
+                 defaults were used. The second element should be an iterable of
+                 Message objects representing any issues which were found when
+                 trying to load configuration - for example, bad values in a
+                 .pylintrc file. It is also possible to simply return None if
+                 neither value is useful.
         """
         # We don't want Pylint to tell us that this method should be a function,
         # it's an "abstract" class after all.
         # pylint: disable=no-self-use
-        return None
+        return None, None
 
     def run(self, found_files):
         """