+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Prospector Changelog
 
 ## Version 0.11.2
 * Fixed a file encoding detection issue when running under Python3
+* If a pylint plugin is specified in a .pylintrc file which cannot be loaded, prospector will now carry on with a warning rather than simply crash
 
 ## Version 0.11.1
 * [#147](https://github.com/landscapeio/prospector/issues/147) Fixed crash when trying to load pylint configuration files in pylint 1.5